Save Luat

+++ SeldatUnilever_Ver1.02/MainWindow.xaml.cs
@@ -192,13 +192,13 @@ namespace SeldatUnilever_Ver1._02
 
             ChartInfo _temp_WT = new ChartInfo();
             _temp_WT.name = "Woking Mode";
-            _temp_WT.value =(int) TotalWorkingTime()+600;
+            _temp_WT.value =(int) (TotalWorkingTime()+600);
             _temp_WT.color = Colors.LightSalmon;
             listRealChartTime.Add(_temp_WT);
 
             ChartInfo _temp_RT = new ChartInfo();
             _temp_RT.name = "Ready Mode";
-            _temp_RT.value = readyTime;
+            _temp_RT.value =(int) readyTime;
             _temp_RT.color = Colors.DarkGray;
             listRealChartTime.Add(_temp_RT);
